third_party/WebKit/LayoutTests/fast/css/pseudo-enabled-disabled-output-element.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<output id="output1"></output>
<output id="output2" disabled></output>
<script>
test(function() {
  o1 = document.getElementById("output1");
  o2 = document.getElementById("output2");
  assert_false(o1.matches(":enabled"));
  assert_false(o2.matches(":disabled"));
}, "This test performs a check for output element not to be considered for :enabled and :disabled CSS selectors.");
</script>
</body>
</html>
